Add a Tiled map renderer and component
This got a bit awkward because map renderers from LibGDX are 1-1 with
maps, so we cannot reuse a renderer for all maps. (Really there should
only be one, but Artemis forces us to divorce the map and its renderer).
Thus, the map component has a renderer. It uses the global SpriteBatch.